fix:  check if snap and nosnap exists and what to do in install_snap, should be fully finished now.

+++ src/edwh_multipass_plugin/tasks.py
@@ -25,25 +25,19 @@ MULTIPASS = "/snap/bin/multipass"
 EW_MP_CONFIG = "~/.config/edwh/multipass.toml"
 DEFAULT_MACHINE_NAME = "dockers"
 
-@task(pre=[edwh.tasks.require_sudo])
-def check_file(c, path="", quiet=False):
-    try:
-        c.sudo(path)
-        return True
-    except FileNotFoundError:
-        if not quiet: print(f"File {path} not found")
-        return False
-
 @task(name="getsnap")
 def install_snap(c, quiet=False):
+    if edwh.tasks.is_installed(c, "snap"):
+        if not quiet: print("Snap already installed")
+        return
     try:
         c.run("rm -f /etc/apt/preferences.d/nosnap.pref")
         if not quiet: print("Nosnap found and removed.")
         c.run("apt update")
     except:
-        if not quiet: print("no nonsnap.pref found")
-    if not quiet: print("installing snap...")
-    # c.sudo("apt install snapd")
+        if not quiet: print("No nonsnap.pref found")
+    if not quiet: print("Installing snap...")
+    c.sudo("apt install snapd")
 
 @task(name="install", pre=[edwh.tasks.require_sudo])
 def install_multipass(c: Connection) -> None:
